requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
psycopg2-binary
redis
django-redis
celery
pillow
requests
python-decouple
drf-spectacular
